JBM-505 decode the email address to get back an valid email
Because of the automatical parameters filter applied to every controller,
the email of EditUserCommand isn't exceptional. By replacing '&#64;' character
as the encoded version of the @ symbol with the original @ one, the email address
will be recovered as what it has been provided and we don't disable the other
encodeded parts.