JBM-519 fix the issue: don't show the spinner
When clicking Next button from the step 1, the spinner should be shown like
uploading files. However, I spent a lot of my time to figure out what is preventing
from the system displaying the spinner. Eventually, I found that the AJAX call to
validate the uploading files was not asynchronous.

If we make an AJAX request asynchronously, some results are lately computed.
Therefore, updating model info for the form after pressing on Next button probably
cause an error. That's why I get rid of the updateModelInfoForm function but call  it
in the success bloc of the AJAX request so that the returned results are already ready.