JBM-519 enhance guessing the model name, description and modelling approach
Originally, the extraction of the model name and description are performed via reading the main file.
In other words, the specific library created and built is used to handle a format and to extract name
and description of the model.

However, when it cannot detect any name and description, it is better to retrieve persistent values
to be saved in the database instead of leaving them empty.

Actually, the fallback values have been initialised but haven't been used anywhere. This commit is
to restructure the logical condition to cover the edge case in which these fields are not populated.

Also, resolved the request JBM-562.