JBM-519 make sure get the correct response once the AJAX call completely succeeds
If we don't handle this work, we will probably end up the situation where the AJAX call
hasn't been complete yet but the app had already hit to the updateModelInfoForm method
to update newly loaded values. Consequently, the variables and objects being referenced
in this method are undefined stopping the system from functioning.