JBM-445 use a flag to retain some model metadata
In reality, we have many situations in which all model metadata of the new revision just need
to be inherited from the former one. This flag enables us not to overwrite the metadata to
ease the creation of ModelTransportCommand object for the RevisionTransportCommand
instance which is the required parameter of the addRevision method.