JBM-519 replace the value of the ajax URL
The health check status endpoint is deliberately designed for checking the status
of BioModels and is used for load balancer and redirection traffic from Hinxton to
Harlow in case of power outage. We might restrict access to this end point for
security concern, therefore, the submission flow has failed recently.

By defining two dummy actions in SubmissionController to replace the health check
status endpoint, we are pretty sure there is no restriction on accessing the
submission flow.