JBM-505 revert unexpectedly encoded user inputs
The automatic encoding user inputs causes, for example,
the acrobat character in the email address, a big problem.
It is that some methods return nothing. For example:

User.findByUsername(username), which the usernae is an email address.

The encoding of user inputs convert '@' so that it doesn't match
with the original value stored in the database.