JBM-566 resolved compiling errors at running tests in IntelliJ IDEA
In the biomodels dom plugin, we use the WcmContent class of the weceem plugin.
Missing its declaration in the BuildConfig file prevents us from running tests
in IntelliJ IDEA as well as under the biomodels dom plugin.

This commit does not directly relate to JBM-566 but it helps to run tests
in IntelliJ IDEA.