JBM-568 handle exceptions causing by downloading in a proper way
This commit includes the solution for the following error:

java.lang.IllegalStateException: getOutputStream() has already been called for this response

The error occurs because we try to access the output stream object after calling include() or
forward() on the response object. Using a local variable to store the input stream object
before populating it to the response's output stream will prevent the error from happening.