JBM-574 use the static-assets path defined in the biomodels-jummp-ingress
By using this path, we are not dependent on the redirection rules defined in the Traffic Manager
which is managed by Web Prod team. Inside the biomodels jummp ingress controller, we are entirely
allowed to defined any other paths to tell which path and service should be associated.

For example:
- /biomodels <-- BioModels Jummp Service
- /biomodels/courses <-- BioModels Static Content <-- BioModels Static Service
- /biomodels/static-assets/ <-- BioModels Static Content <-- BioModels Static Service

This comment is related to the ticket JBM-574 as defining a service in BioModels K8s cluster
to serve static contents will move away of VMs. Historically, the static contents are located in
shared NFS partitions which Tomcat servers can access from VMs.

Revise the commits linked to the ticket BMN-68 in the repo nginx-static-content.