JBM-575 create associations for revision, user and contribution role
In this commit, I have made the changes on adding the current user
as the contributor for the updating revision to the database and
fetching contributors from the database instead of inferring from the
revision owner.

Adding a user as the revision contributor is one of the post submission
processes.

TODO: reconcile the contributors and createrUsernames states.