JBM-519 update the list of changes made instead of overwriting it
The list of changes made must be updated (i.e. added or removed) when we
go back and forth of the submission flow. To avoid re-adding the messages,
we replaced the data structure of the changesMade variable from List to Set.
At the step of doing the last check on the submission data, the changesMade
is passed to the back end (i.e. controller) to perform validation, so we have also
to convert back from Set to List to ensure the compatibility of data structure.