JBM-519 make sure all checks done and completed before showing the result message
When disabling asynchronous calls, we want to make sure that all checks and persistence are done
before the completion message is displayed to the submitter. This change could lead to a delay when
the model files are too large. However, it makes the steps consistent  in the fact that the backend services are successfully completed and the corresponding message is displayed.