JBM-572 reimplement downloading models in CombineArchive
Regarding large models, generating CombineArchive file and streaming
it to clients on the fly causes RunOutOfMemory exception. By limiting
300 MB on the total size of the model files, we can serve out clients
the CombineArchive file in the fly or the one already created and left
in the model directory on EBI FTP public or a temp directory that will
be expired within 1 hour, for example.