JBM-468 implement this service to fix vcs commit hashes
When trying to load a non-latest revision, I found that the commit hashes saved in
the db are completely different from the ones in the git-based model directory.

To fix that bug, we will copy and match revisions with the commit hashes. It is a tedious
and error-prone job. Now, it could be fixed by loading the given model and calling
vcsService.fixVcsIds.