JBM-585 resolved ClassNotFoundException
The removal of the weceem plugin could impact all Filters classes.
However, moving all Filters classes to the conf directory is better
than leaving them in packages. According to the documentation, we
created them incorrectly.

We will work out on how to put them back in separate packages.

The related ticket is JBM-613.