JBM-519 add an optional argument as the map of the submission data
The aim is to use this map to store some info where we need to use during
the post submission, for example. This change also makes sure that the former
calls will continue working because the default value to pass to this optional
argument is null.

One of the requests is to add or change the contributor role during
the submission or update process done via REST API (see JBM-325).