JBM-608 expose a REST API EP for generating the OMEX file of models
I tried to tackle JBM-608 but it failed. The reasons are:
i)  The limit size of the concurrent queue in OmexService is 10
ii) Some models take too long to complete the generation of OMEX file.
iii)The work has to connect to OmexService directly without touching
    BioModels service.

When this work is done, JBM-660 is probably easier.