JBM-468 get the first and last revision bypassing ACLs check
This method is implemented to allow improving core services.
When bypassing ACLs checks, this method will be used by admins.
The admins sometimes need to manage data such data corrections
or data retrievals.

This commit is connected to JBM-647 when retrieving and caching annotations.

When working with JBM-671, I found it is better to cache the annotation
accessions on Redis server. It makes easier to deal with annotations.