JBM-663 make sure to load all contributors of the given model
When users don't indicate a revision number, the latest revision will be loaded.
The contributors now are all the ones who have contributed to all revisions.
Otherwise, the contributors are only the people who have contributed to the revision
in question.
For example: BIOMD0000000706 has 6 revisions as of writing this commit.
* the contributors of BIOMD0000000706.1 are the ones submitted or updated this revision only.
* the contributors of BIOMD0000000706.4 will be ones submitted and updated this revision and
  the previous revisions.
* the contributors of BIOMD0000000706 (without a revision number) are all people.